+++ air/src/execution_step/air/call/prev_result_handler.rs
@@ -50,7 +50,7 @@ pub(super) fn handle_prev_state<'i>(
         // this call was failed on one of the previous executions,
         // here it's needed to bubble this special error up
         CallServiceFailed(ret_code, err_msg) => {
-            exec_ctx.subtree_complete = false;
+            exec_ctx.subgraph_complete = false;
             let ret_code = *ret_code;
             let err_msg = err_msg.clone();
             trace_ctx.meet_call_end(prev_result);
@@ -67,7 +67,7 @@ pub(super) fn handle_prev_state<'i>(
                 }
                 // result hasn't been prepared yet
                 None => {
-                    exec_ctx.subtree_complete = false;
+                    exec_ctx.subgraph_complete = false;
                     Ok(StateDescriptor::not_ready(prev_result))
                 }
             }
@@ -79,7 +79,7 @@ pub(super) fn handle_prev_state<'i>(
                 return Ok(StateDescriptor::can_execute_now(prev_result));
             }
 
-            exec_ctx.subtree_complete = false;
+            exec_ctx.subgraph_complete = false;
             Ok(StateDescriptor::cant_execute_now(prev_result))
         }
         // this instruction's been already executed