+++ crates/air-lib/trace-handler/src/state_automata/par_fsm/par_builder.rs
@@ -20,37 +20,37 @@ use super::*;
 #[derive(Debug, Default, Clone)]
 pub(super) struct ParBuilder {
     saved_states_count: usize,
-    left_subtree_size: usize,
-    right_subtree_size: usize,
+    left_subgraph_size: usize,
+    right_subgraph_size: usize,
 }
 
 impl ParBuilder {
     // StateInserter here needs to guaranteed that ParBuilder creates after it,
-    // it must be so to right track a left subtree size
+    // it must be so to right track a left subgraph size
     pub(super) fn from_keeper(data_keeper: &DataKeeper, _: &StateInserter) -> Self {
         let saved_states_count = data_keeper.result_states_count();
 
         Self {
             saved_states_count,
-            left_subtree_size: 0,
-            right_subtree_size: 0,
+            left_subgraph_size: 0,
+            right_subgraph_size: 0,
         }
     }
 
-    pub(super) fn track(&mut self, data_keeper: &DataKeeper, subtree_type: SubtreeType) {
+    pub(super) fn track(&mut self, data_keeper: &DataKeeper, subgraph_type: SubgraphType) {
         let prev_states_count = self.saved_states_count;
         let states_count = data_keeper.result_states_count();
         let resulted_states_count = states_count - prev_states_count;
 
-        match subtree_type {
-            SubtreeType::Left => self.left_subtree_size = resulted_states_count,
-            SubtreeType::Right => self.right_subtree_size = resulted_states_count,
+        match subgraph_type {
+            SubgraphType::Left => self.left_subgraph_size = resulted_states_count,
+            SubgraphType::Right => self.right_subgraph_size = resulted_states_count,
         }
         self.saved_states_count = data_keeper.result_trace.len();
     }
 
     pub(super) fn build(self) -> ExecutedState {
         // TODO: check that usize could be converted into u32
-        ExecutedState::par(self.left_subtree_size, self.right_subtree_size)
+        ExecutedState::par(self.left_subgraph_size, self.right_subgraph_size)
     }
 }