+++ crates/air-lib/trace-handler/src/state_automata/par_fsm/state_handler/new_states_calculation.rs
@@ -21,11 +21,11 @@ pub(super) fn compute_new_states(
     data_keeper: &DataKeeper,
     prev_par: ParResult,
     current_par: ParResult,
-    subtree_type: SubtreeType,
+    subgraph_type: SubgraphType,
 ) -> FSMResult<CtxStatesPair> {
-    let (prev_len, current_len) = match subtree_type {
-        SubtreeType::Left => (prev_par.left_size, current_par.left_size),
-        SubtreeType::Right => {
+    let (prev_len, current_len) = match subgraph_type {
+        SubgraphType::Left => (prev_par.left_size, current_par.left_size),
+        SubgraphType::Right => {
             let prev_par_size = prev_par.size().ok_or(StateFSMError::ParLenOverflow(prev_par))?;
             let current_par_size = current_par.size().ok_or(StateFSMError::ParLenOverflow(current_par))?;
 
@@ -40,15 +40,15 @@ pub(super) fn compute_new_states(
     Ok(pair)
 }
 
-fn compute_new_state(par_subtree_len: usize, slider: &TraceSlider, par: ParResult) -> FSMResult<CtxState> {
+fn compute_new_state(par_subgraph_len: usize, slider: &TraceSlider, par: ParResult) -> FSMResult<CtxState> {
     let pos = slider
         .position()
-        .checked_add(par_subtree_len)
+        .checked_add(par_subgraph_len)
         .ok_or_else(|| StateFSMError::ParPosOverflow(par, slider.position(), MergeCtxType::Previous))?;
 
     let subtrace_len = slider
         .subtrace_len()
-        .checked_sub(par_subtree_len)
+        .checked_sub(par_subgraph_len)
         .ok_or_else(|| StateFSMError::ParLenUnderflow(par, slider.subtrace_len(), MergeCtxType::Current))?;
 
     let new_state = CtxState::new(pos, subtrace_len);